requirements.txt
langchain
langchain_ibm
langchain-community
langgraph
gradio
chromadb
ibm-watsonx-ai
pdfminer.six
pypdf
tqdm
numpy
